[Spécification d'histogrammes]  + Algorithme de construction de la fonction de correspondance à partir d'un histogramme cible OK  + Spécification d'un histogramme constant OK  + Spécification d'un histogramme linéaire OK  + Spécification d'un histogramme racine carrée OK

+++ tp1/Graphics/GrayscaleCumulativeHistogram.cpp
@@ -34,7 +34,7 @@ void GrayscaleCumulativeHistogram::reset(int size)
 void GrayscaleCumulativeHistogram::update(const GrayscaleHistogram &sourceHisto)
 {
     int sourceHistoSize;
-    unsigned long currentSum;
+    double currentSum;
     std::vector<double> sourceHistoValues = sourceHisto.intensityProbabilities();
 
     currentSum = 0;
